Facebook comments sorta?
There was hours of research trying to figure out how this magic works.
I'm still quite unsure, but I'll pick up from here later on. Not much to
show. All I know is that the fb:comments (deprecated) actually throws
the comment to facebook, but doesn't post it. The fb-comments is not
making it to facebook (current method). I've been unsuccessfully
fiddling with the permissions and stuff to see why it's not posting when
it should.