Facebook commits working 99.9%
I fixed up the facebook box so that if you tell it to post to facebook,
it will post to your profile like it should. I've also got it directing
to where the 1&1 page is just in case we get around to uploading it and
getting it working. Also, there is an unreachable blahblahblah warning
which I don't think I can get rid of.